New component for Authorization and removing authorization logic from the operation-console
