+++ src/components/operations/operation-details/ko/runtime/operation-console.ts
@@ -3,10 +3,7 @@ import * as validation from "knockout.validation";
 import { ISettingsProvider } from "@paperbits/common/configuration";
 import { HttpClient, HttpRequest, HttpResponse } from "@paperbits/common/http";
 import { Component, OnMounted, Param } from "@paperbits/common/ko/decorators";
-import { SessionManager } from "@paperbits/common/persistence/sessionManager";
 import { RequestBodyType, ServiceSkuName, TypeOfApi } from "../../../../../constants";
-import { SubscriptionState } from "../../../../../contracts/subscription";
-import { UnauthorizedError } from "../../../../../errors/unauthorizedError";
 import { Api } from "../../../../../models/api";
 import { AuthorizationServer } from "../../../../../models/authorizationServer";
 import { ConsoleHeader } from "../../../../../models/console/consoleHeader";
@@ -15,25 +12,17 @@ import { ConsoleParameter } from "../../../../../models/console/consoleParameter
 import { KnownHttpHeaders } from "../../../../../models/knownHttpHeaders";
 import { KnownStatusCodes } from "../../../../../models/knownStatusCodes";
 import { Operation } from "../../../../../models/operation";
-import { Product } from "../../../../../models/product";
 import { Revision } from "../../../../../models/revision";
 import { RouteHelper } from "../../../../../routing/routeHelper";
 import { ApiService } from "../../../../../services/apiService";
-import { OAuthService } from "../../../../../services/oauthService";
-import { ProductService } from "../../../../../services/productService";
 import { TemplatingService } from "../../../../../services/templatingService";
 import { TenantService } from "../../../../../services/tenantService";
-import { UsersService } from "../../../../../services/usersService";
 import { Utils } from "../../../../../utils";
-import { GrantTypes } from "./../../../../../constants";
-import { OAuthSession, StoredCredentials } from "./oauthSession";
 import template from "./operation-console.html";
 import { ResponsePackage } from "./responsePackage";
 import { templates } from "./templates/templates";
 import { LogItem, WebsocketClient } from "./websocketClient";
 import { KnownMimeTypes } from "../../../../../models/knownMimeTypes";
-
-const oauthSessionKey = "oauthSession";
 
 @Component({
     selector: "operation-console",
@@ -48,22 +37,13 @@ export class OperationConsole {
     public readonly responseStatusText: ko.Observable<string>;
     public readonly responseBody: ko.Observable<string>;
     public readonly responseHeadersString: ko.Observable<string>;
-    public readonly products: ko.Observable<Product[]>;
-    public readonly selectedSubscriptionKey: ko.Observable<string>;
-    public readonly subscriptionKeyRequired: ko.Observable<boolean>;
     public readonly selectedLanguage: ko.Observable<string>;
-    public readonly selectedProduct: ko.Observable<Product>;
     public readonly requestError: ko.Observable<string>;
     public readonly codeSample: ko.Observable<string>;
     public readonly selectedHostname: ko.Observable<string>;
     public readonly isHostnameWildcarded: ko.Computed<boolean>;
     public readonly hostnameSelectionEnabled: ko.Observable<boolean>;
     public readonly wildcardSegment: ko.Observable<string>;
-    public readonly selectedGrantType: ko.Observable<string>;
-    public readonly username: ko.Observable<string>;
-    public readonly password: ko.Observable<string>;
-    public readonly authorizationError: ko.Observable<string>;
-    public readonly authenticated: ko.Observable<boolean>;
     public isConsumptionMode: boolean;
     public templates: Object;
     public backendUrl: string;
@@ -81,16 +61,11 @@ export class OperationConsole {
     constructor(
         private readonly apiService: ApiService,
         private readonly tenantService: TenantService,
-        private readonly usersService: UsersService,
-        private readonly productService: ProductService,
         private readonly httpClient: HttpClient,
         private readonly routeHelper: RouteHelper,
-        private readonly oauthService: OAuthService,
-        private readonly sessionManager: SessionManager,
         private readonly settingsProvider: ISettingsProvider
     ) {
         this.templates = templates;
-        this.products = ko.observable();
 
         this.requestError = ko.observable();
         this.responseStatusCode = ko.observable();
@@ -104,22 +79,14 @@ export class OperationConsole {
         this.hostnames = ko.observable();
         this.consoleOperation = ko.observable();
         this.secretsRevealed = ko.observable();
-        this.selectedSubscriptionKey = ko.observable();
-        this.subscriptionKeyRequired = ko.observable();
         this.working = ko.observable(true);
         this.sendingRequest = ko.observable(false);
         this.codeSample = ko.observable();
-        this.selectedProduct = ko.observable();
         this.onFileSelect = this.onFileSelect.bind(this);
         this.selectedHostname = ko.observable("");
         this.hostnameSelectionEnabled = ko.observable();
         this.isHostnameWildcarded = ko.computed(() => this.selectedHostname().includes("*"));
-        this.selectedGrantType = ko.observable();
         this.authorizationServer = ko.observable();
-        this.username = ko.observable();
-        this.password = ko.observable();
-        this.authorizationError = ko.observable();
-        this.authenticated = ko.observable(false);
 
         this.useCorsProxy = ko.observable(false);
         this.wildcardSegment = ko.observable();
@@ -182,11 +149,9 @@ export class OperationConsole {
 
             this.setHostname(hostname);
         });
-        this.selectedSubscriptionKey.subscribe(this.applySubscriptionKey.bind(this));
         this.api.subscribe(this.resetConsole);
         this.operation.subscribe(this.resetConsole);
         this.selectedLanguage.subscribe(this.updateRequestSummary);
-        this.selectedGrantType.subscribe(this.onGrantTypeChange);
     }
 
     private async resetConsole(): Promise<void> {
@@ -205,9 +170,6 @@ export class OperationConsole {
         this.responseStatusCode(null);
         this.responseStatusText(null);
         this.responseBody(null);
-        this.selectedSubscriptionKey(null);
-        this.subscriptionKeyRequired(!!selectedApi.subscriptionRequired);
-        this.selectedProduct(null);
 
         const operation = await this.apiService.getOperation(selectedOperation.id);
         const consoleOperation = new ConsoleOperation(selectedApi, operation);
@@ -236,12 +198,6 @@ export class OperationConsole {
 
         if (this.api().apiVersionSet && this.api().apiVersionSet.versioningScheme === "Header") {
             this.setVersionHeader();
-        }
-
-        await this.setupOAuth();
-
-        if (selectedApi.subscriptionRequired) {
-            await this.loadSubscriptionKeys();
         }
 
         this.updateRequestSummary();
@@ -274,52 +230,6 @@ export class OperationConsole {
         }
 
         consoleOperation.urlTemplate = "";
-    }
-
-    private async loadSubscriptionKeys(): Promise<void> {
-        const userId = await this.usersService.getCurrentUserId();
-
-        if (!userId) {
-            return;
-        }
-
-        const pageOfProducts = await this.apiService.getAllApiProducts(this.api().id);
-        const products = pageOfProducts && pageOfProducts.value ? pageOfProducts.value : [];
-        const pageOfSubscriptions = await this.productService.getSubscriptions(userId);
-        const subscriptions = pageOfSubscriptions.value.filter(subscription => subscription.state === SubscriptionState.active);
-        const availableProducts = [];
-
-        products.forEach(product => {
-            const keys = [];
-
-            subscriptions.forEach(subscription => {
-                if (!this.productService.isScopeSuitable(subscription.scope, this.api().name, product.name)) {
-                    return;
-                }
-
-                keys.push({
-                    name: `Primary: ${subscription.name?.trim() || subscription.primaryKey.substr(0, 4)}`,
-                    value: subscription.primaryKey
-                });
-
-                keys.push({
-                    name: `Secondary: ${subscription.name?.trim() || subscription.secondaryKey.substr(0, 4)}`,
-                    value: subscription.secondaryKey
-                });
-            });
-
-            if (keys.length > 0) {
-                availableProducts.push({ name: product.displayName, subscriptionKeys: keys });
-            }
-        });
-
-        this.products(availableProducts);
-
-        if (availableProducts.length > 0) {
-            const subscriptionKey = availableProducts[0].subscriptionKeys[0].value;
-            this.selectedSubscriptionKey(subscriptionKey);
-            this.applySubscriptionKey(subscriptionKey);
-        }
     }
 
     private loadRequestLanguagesRest(): Object[] {
@@ -359,14 +269,6 @@ export class OperationConsole {
         this.updateRequestSummary();
     }
 
-    private findHeader(name: string): ConsoleHeader {
-        const searchName = name.toLocaleLowerCase();
-
-        return this.consoleOperation().request
-            .headers()
-            .find(x => x.name()?.toLocaleLowerCase() === searchName);
-    }
-
     public addQueryParameter(): void {
         this.consoleOperation().request.queryParameters.push(new ConsoleParameter());
         this.updateRequestSummary();
@@ -377,129 +279,12 @@ export class OperationConsole {
         this.updateRequestSummary();
     }
 
-    private applySubscriptionKey(subscriptionKey: string): void {
-        if (!this.consoleOperation()) {
-            return;
-        }
-
-        if (this.api().type === TypeOfApi.webSocket) {
-            this.setSubscriptionKeyParameter(subscriptionKey)
-        } else {
-            this.setSubscriptionKeyHeader(subscriptionKey);
-        }
-        this.updateRequestSummary();
-    }
-
     private setNoCacheHeader(): void {
         this.consoleOperation().setHeader(KnownHttpHeaders.CacheControl, "no-cache", "string", "Disable caching.");
     }
 
     private setVersionHeader(): void {
         this.consoleOperation().setHeader(this.api().apiVersionSet.versionHeaderName, this.api().apiVersion, "string", "API version");
-    }
-
-    private getSubscriptionKeyHeaderName(): string {
-        let subscriptionKeyHeaderName: string = KnownHttpHeaders.OcpApimSubscriptionKey;
-
-        if (this.api().subscriptionKeyParameterNames && this.api().subscriptionKeyParameterNames.header) {
-            subscriptionKeyHeaderName = this.api().subscriptionKeyParameterNames.header;
-        }
-
-        return subscriptionKeyHeaderName;
-    }
-
-    private getSubscriptionKeyParamName(): string {
-        let subscriptionKeyParamName = "subscription-key";
-
-        if (this.api().subscriptionKeyParameterNames && this.api().subscriptionKeyParameterNames.query) {
-            subscriptionKeyParamName = this.api().subscriptionKeyParameterNames.query;
-        }
-
-        return subscriptionKeyParamName;
-    }
-
-    private getSubscriptionKeyParam(): ConsoleParameter {
-        const subscriptionKeyParamName = this.getSubscriptionKeyParamName();
-        const searchName = subscriptionKeyParamName.toLocaleLowerCase();
-
-        return this.consoleOperation().request.queryParameters().find(x => x.name()?.toLocaleLowerCase() === searchName);
-    }
-
-    private setSubscriptionKeyParameter(subscriptionKey: string): void {
-        const subscriptionKeyParam = this.getSubscriptionKeyParam();
-        this.removeQueryParameter(subscriptionKeyParam);
-
-        if (!subscriptionKey) {
-            return;
-        }
-
-        const subscriptionKeyParamName = this.getSubscriptionKeyParamName();
-
-        const keyParameter = new ConsoleParameter();
-        keyParameter.name(subscriptionKeyParamName);
-        keyParameter.value(subscriptionKey);
-        keyParameter.secret = true;
-        keyParameter.type = "string";
-        keyParameter.canRename = false;
-        keyParameter.required = true;
-
-        this.consoleOperation().request.queryParameters.push(keyParameter);
-        this.updateRequestSummary();
-    }
-
-    private getSubscriptionKeyHeader(): ConsoleHeader {
-        const subscriptionKeyHeaderName = this.getSubscriptionKeyHeaderName();
-        return this.findHeader(subscriptionKeyHeaderName);
-    }
-
-    private setSubscriptionKeyHeader(subscriptionKey: string): void {
-        this.removeSubscriptionKeyHeader();
-
-        if (!subscriptionKey) {
-            return;
-        }
-
-        const subscriptionKeyHeaderName = this.getSubscriptionKeyHeaderName();
-
-        const keyHeader = new ConsoleHeader();
-        keyHeader.name(subscriptionKeyHeaderName);
-        keyHeader.value(subscriptionKey);
-        keyHeader.description = "Subscription key.";
-        keyHeader.secret = true;
-        keyHeader.inputTypeValue = "password";
-        keyHeader.type = "string";
-        keyHeader.required = true;
-
-        this.consoleOperation().request.headers.push(keyHeader);
-        this.updateRequestSummary();
-    }
-
-    private removeAuthorizationHeader(): void {
-        const authorizationHeader = this.findHeader(KnownHttpHeaders.Authorization);
-        this.removeHeader(authorizationHeader);
-        this.authenticated(false);
-    }
-
-    private setAuthorizationHeader(accessToken: string): void {
-        this.removeAuthorizationHeader();
-
-        const keyHeader = new ConsoleHeader();
-        keyHeader.name(KnownHttpHeaders.Authorization);
-        keyHeader.value(accessToken);
-        keyHeader.description = "Subscription key.";
-        keyHeader.secret = true;
-        keyHeader.inputTypeValue = "password";
-        keyHeader.type = "string";
-        keyHeader.required = true;
-
-        this.consoleOperation().request.headers.push(keyHeader);
-        this.updateRequestSummary();
-        this.authenticated(true);
-    }
-
-    private removeSubscriptionKeyHeader(): void {
-        const subscriptionKeyHeader = this.getSubscriptionKeyHeader();
-        this.removeHeader(subscriptionKeyHeader);
     }
 
     public async updateRequestSummary(): Promise<void> {
@@ -756,128 +541,4 @@ export class OperationConsole {
         return this.routeHelper.getApiReferenceUrl(this.api().name);
     }
 
-    private getSessionRecordKey(authorizationServerName: string, scopeOverride: string): string {
-        let recordKey = authorizationServerName;
-
-        if (scopeOverride) {
-            recordKey += `-${scopeOverride}`;
-        }
-
-        return recordKey;
-    }
-
-    private async setupOAuth(): Promise<void> {
-        const authorizationServer = this.authorizationServer();
-
-        if (!authorizationServer) {
-            this.selectedGrantType(null);
-            return;
-        }
-
-        const api = this.api();
-        const scopeOverride = api.authenticationSettings?.oAuth2?.scope;
-        const storedCredentials = await this.getStoredCredentials(authorizationServer.name, scopeOverride);
-
-        if (storedCredentials) {
-            this.selectedGrantType(storedCredentials.grantType);
-            this.setAuthorizationHeader(storedCredentials.accessToken);
-        }
-    }
-
-    private async getStoredCredentials(serverName: string, scopeOverride: string): Promise<StoredCredentials> {
-        const oauthSession = await this.sessionManager.getItem<OAuthSession>(oauthSessionKey);
-        const recordKey = this.getSessionRecordKey(serverName, scopeOverride);
-        const storedCredentials = oauthSession?.[recordKey];
-
-        try {
-            /* Trying to check if it's a JWT token and, if yes, whether it got expired. */
-            const jwtToken = Utils.parseJwt(storedCredentials.accessToken.replace(/^bearer /i, ""));
-            const now = new Date();
-
-            if (now > jwtToken.exp) {
-                await this.clearStoredCredentials();
-                return null;
-            }
-        }
-        catch (error) {
-            // do nothing
-        }
-
-        return storedCredentials;
-    }
-
-    private async setStoredCredentials(serverName: string, scopeOverride: string, grantType: string, accessToken: string): Promise<void> {
-        const oauthSession = await this.sessionManager.getItem<OAuthSession>(oauthSessionKey) || {};
-        const recordKey = this.getSessionRecordKey(serverName, scopeOverride);
-
-        oauthSession[recordKey] = {
-            grantType: grantType,
-            accessToken: accessToken
-        };
-
-        await this.sessionManager.setItem<object>(oauthSessionKey, oauthSession);
-    }
-
-    private async clearStoredCredentials(): Promise<void> {
-        await this.sessionManager.removeItem(oauthSessionKey);
-        this.removeAuthorizationHeader();
-    }
-
-    public async authenticateOAuthWithPassword(): Promise<void> {
-        try {
-            this.authorizationError(null);
-
-            const api = this.api();
-            const authorizationServer = this.authorizationServer();
-            const scopeOverride = api.authenticationSettings?.oAuth2?.scope;
-            const serverName = authorizationServer.name;
-
-            if (scopeOverride) {
-                authorizationServer.scopes = [scopeOverride];
-            }
-
-            const accessToken = await this.oauthService.authenticatePassword(this.username(), this.password(), authorizationServer);
-            await this.setStoredCredentials(serverName, scopeOverride, GrantTypes.password, accessToken);
-
-            this.setAuthorizationHeader(accessToken);
-        }
-        catch (error) {
-            if (error instanceof UnauthorizedError) {
-                this.authorizationError(error.message);
-                return;
-            }
-
-            this.authorizationError("Oops, something went wrong. Try again later.");
-        }
-    }
-
-    private async onGrantTypeChange(grantType: string): Promise<void> {
-        await this.clearStoredCredentials();
-
-        if (!grantType || grantType === GrantTypes.password) {
-            return;
-        }
-
-        await this.authenticateOAuth(grantType);
-    }
-
-    /**
-     * Initiates specified authentication flow.
-     * @param grantType OAuth grant type, e.g. "implicit" or "authorization_code".
-     */
-    public async authenticateOAuth(grantType: string): Promise<void> {
-        const api = this.api();
-        const authorizationServer = this.authorizationServer();
-        const scopeOverride = api.authenticationSettings?.oAuth2?.scope;
-        const serverName = authorizationServer.name;
-
-        if (scopeOverride) {
-            authorizationServer.scopes = [scopeOverride];
-        }
-
-        const accessToken = await this.oauthService.authenticate(grantType, authorizationServer);
-        await this.setStoredCredentials(serverName, scopeOverride, grantType, accessToken);
-
-        this.setAuthorizationHeader(accessToken);
-    }
 }
