+++ src/config.publish.json
@@ -1,6 +1,6 @@
 {
     "environment": "publishing",
-    "managementApiUrl": "https://<service-name>.management.azure-api.net",
-    "managementApiAccessToken": "SharedAccessSignature...",
+    "managementApiUrl": "https://jbtests-apimanagement.management.azure-api.net",
+    "managementApiAccessToken": "SharedAccessSignature integration&202111211609&2U/mu9/5i8T2kd85FrBFPMooM6Akr/QcgK2V0q1vDAgJDGRKxLaLzEFCdwkhK/mGQiR5hK31LUkvstMnIauYbw==",
     "useHipCaptcha": false
 }
